Fix duration on some configurations
As the time spots are saved as floating point values and rounded on
periodic values we have to correct the initial assumption that values
can be trusted and as is translated into seconds.

To do this we simply attach the difference between the total time and
the expected day length to the last transition in the queue.